Fix build for boot2qt eAndroid
Reverts the removal of Android support in WTF, and brings the support
up to date.

Merged change to MacroAssemblerARM.cpp from QtScript.

Replaces use of statvfs with statfs in one place.

Replaces shm-based shared memory with ashmem-based in WebKit2.

+++ Source/WebKit/qt/WebCoreSupport/PlatformStrategiesQt.cpp
@@ -169,7 +169,7 @@ void PlatformStrategiesQt::getPluginInfo(const WebCore::Page* page, Vector<WebCo
     PluginDatabase* db = PluginDatabase::installedPlugins();
     const Vector<PluginPackage*> &plugins = db->plugins();
 
-    for (int i = 0; i < plugins.size(); ++i) {
+    for (size_t i = 0; i < plugins.size(); ++i) {
         PluginInfo info;
         PluginPackage* package = plugins[i];
